// Online-Store/src/components/types/enums.ts
enum Color {
  white = "white",
  purple = "purple",
  blue = "blue",
  black = "black",
}
enum Manufacturer {
  xiaomi = "xiaomi",
  apple = "apple",
  samsung = "samsung",
}
enum NumberOfCameras {
  camera1 = "camera1",
  camera2 = "camera2",
  camera3 = "camera3",
}
enum ClassNameActive {
  activeCheckbox = "active-checkbox",
  activeIcon = "active-icon",
}
enum sortAscendDescend {
  AZ = "A-Z",
  ZA = "Z-A",
  min = "2017-2022",
  max = "2022-2017",
}
export {
  Color,
  Manufacturer,
  NumberOfCameras,
  ClassNameActive,
  sortAscendDescend,
};